fix network

+++ Sources/Server/Network/User/tcpUser.cpp
@@ -27,7 +27,7 @@ void rtype::tcpUser::doRead(const std::error_code &ec, size_t bytes)
     (void)bytes;
     if (!ec)
     {
-        std::cout << _data[0] << std::endl;
+        std::cout << (int)_data[0] << std::endl;
         read();
         _sizeInput++;
     }
